requirements.txt
opencv-python-headless
numpy
matplotlib
scikit-image
scipy
seaborn
pycryptodome
reedsolo
